# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.OrderStatus:
    enum:
    - pending
    - paid
    - shipped
    - completed
    - cancelled
    type: string
    x-enum-comments:
      OrderStatusCancelled: 已取消
      OrderStatusCompleted: 已完成
      OrderStatusPaid: 已支付
      OrderStatusPending: 待支付
      OrderStatusShipped: 已发货
    x-enum-descriptions:
    - 待支付
    - 已支付
    - 已发货
    - 已完成
    - 已取消
    x-enum-varnames:
    - OrderStatusPending
    - OrderStatusPaid
    - OrderStatusShipped
    - OrderStatusCompleted
    - OrderStatusCancelled
...
  service.CreateOrderRequest:
    properties:
      cart_item_ids:
        items:
          type: integer
        minItems: 1
        type: array
    required:
    - cart_item_ids
    type: object
  service.CreateProductRequest:
    properties:
      base_price:
        type: number
      brand:
        type: string
      category:
        type: string
      description:
        type: string
      material:
        type: string
      name:
        type: string
    required:
    - base_price
    - name
    type: object
  service.OrderItemResponse:
    properties:
      color:
        type: string
      design_image_url:
        type: string
      id:
        type: integer
      price:
        type: number
      product_name:
        type: string
      quantity:
        type: integer
      size:
        type: string
    type: object
  service.OrderResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      order_items:
        items:
          $ref: '#/definitions/service.OrderItemResponse'
        type: array
      order_sn:
        type: string
      status:
        $ref: '#/definitions/model.OrderStatus'
      total_amount:
        type: number
    type: object
  service.ProductResponse:
    properties:
      base_price:
        type: number
      brand:
        type: string
      category:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      material:
        type: string
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: AI驱动的T恤设计商店API服务，提供用户认证、AI设计生成等功能
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: AI T-Shirt Shop API
  version: "1.0"
paths:
...
  /orders:
    get:
      consumes:
      - application/json
      description: 获取当前用户的所有订单
      produces:
      - application/json
      responses:
        "200":
          description: 订单列表
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 内部服务器错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取用户订单列表
      tags:
      - orders
    post:
      consumes:
      - application/json
      description: 从购物车创建订单，将购物车中的商品转换为订单
      parameters:
      - description: 创建订单请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 订单创建成功
          schema:
            $ref: '#/definitions/service.OrderResponse'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 内部服务器错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 创建订单
      tags:
      - orders
  /orders/{id}:
    get:
      consumes:
      - application/json
      description: 根据订单ID获取订单详细信息
      parameters:
      - description: 订单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 订单详情
          schema:
            $ref: '#/definitions/service.OrderResponse'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 订单不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 内部服务器错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 根据ID获取订单
      tags:
      - orders
  /orders/{id}/status:
    put:
      consumes:
      - application/json
      description: 更新订单状态（支付、发货、完成等）
      parameters:
      - description: 订单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 状态更新请求参数
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 状态更新成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 订单不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 内部服务器错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 更新订单状态
      tags:
      - orders
  /orders/sn/{order_sn}:
    get:
      consumes:
      - application/json
      description: 根据订单号获取订单详细信息
      parameters:
      - description: 订单号
        in: path
        name: order_sn
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 订单详情
          schema:
            $ref: '#/definitions/service.OrderResponse'
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 订单不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 内部服务器错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 根据订单号获取订单
      tags:
      - orders
  /products:
    get:
      consumes:
      - application/json
      description: 获取所有可用的T恤商品模板
      produces:
      - application/json
      responses:
        "200":
          description: 商品列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 内部服务器错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取所有商品
      tags:
      - products
    post:
      consumes:
      - application/json
      description: 创建新的T恤商品模板
      parameters:
      - description: 创建商品请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 商品创建成功
          schema:
            $ref: '#/definitions/service.ProductResponse'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 内部服务器错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 创建商品
      tags:
      - products
  /products/{id}:
    get:
      consumes:
      - application/json
      description: 根据商品ID获取特定的T恤商品模板信息
      parameters:
      - description: 商品ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 商品信息
          schema:
            $ref: '#/definitions/service.ProductResponse'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 商品不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 内部服务器错误
          schema:
            additionalProperties: true
            type: object
      summary: 根据ID获取商品
      tags:
      - products
  /products/category:
    get:
      consumes:
      - application/json
      description: 根据商品分类获取T恤商品列表
      parameters:
      - description: 商品分类
        in: query
        name: category
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 商品列表
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 内部服务器错误
          schema:
            additionalProperties: true
            type: object
      summary: 根据分类获取商品
      tags:
      - products
...
securityDefinitions:
  BearerAuth:
    description: Bearer token for API authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"